requirements.txt
Flask==2.3.3
Pillow==10.0.1
numpy==1.24.3
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
requests>=2.28.0
typing-extensions>=4.10.0